adding some base stuff

+++ src/database/DAO/base-dao.js
@@ -1,9 +1,12 @@
 import MongoConnectionManager from "../mongo-connection-manager";
 
 const ObjectId = require("mongodb").ObjectId;
-
+//TODO is there a way to abstract the collection checks?
 const BaseDao = {
     getAll: async (collection, obj, exclusion) => {
+        if (typeof collection !== "string") {
+            throw new Error("Collection either not provided or is not a string");
+        }
         const db = await MongoConnectionManager.getConnection();
         obj = obj._id ? {
             ...obj,
@@ -22,7 +25,9 @@ const BaseDao = {
         });
     },
     getBy: async (collection, obj) => {
-
+        if (typeof collection !== "string") {
+            throw new Error("Collection either not provided or is not a string");
+        }
         return new Promise(async (resolve, reject) => {
             const db = await MongoConnectionManager.getConnection();
             obj = obj._id ? {
@@ -30,6 +35,7 @@ const BaseDao = {
                 _id: new ObjectId(obj._id)
             } : obj;
             db.collection(collection).findOne(obj, (err, result) => {
+                    console.log(err);
                     if (err) {
                         reject("Couldn't findOne");
                     } else {
@@ -43,9 +49,12 @@ const BaseDao = {
         });
 
     },
-    create: async (collection, obj) => {
-        const db = await MongoConnectionManager.getConnection();
-        return new Promise((resolve, reject) => {
+    create: (collection, obj) => {
+        if (typeof collection !== "string") {
+            throw new Error("Collection either not provided or is not a string");
+        }
+        return new Promise(async (resolve, reject) => {
+            const db = await MongoConnectionManager.getConnection();
             db.collection(collection).insertOne(obj, (err, res) => {
                 if (err) {
                     reject(err);
@@ -57,6 +66,9 @@ const BaseDao = {
         });
     },
     updateById: async (collection, obj) => {
+        if (typeof collection !== "string") {
+            throw new Error("Collection either not provided or is not a string");
+        }
         const storedObj = await BaseDao.getBy(collection, {_id: obj._id});
         await BaseDao.removeById(collection, obj._id);
         const o = {
@@ -68,6 +80,9 @@ const BaseDao = {
         await BaseDao.create(collection, o);
     },
     removeById: async (collection, _id) => {
+        if (typeof collection !== "string") {
+            throw new Error("Collection either not provided or is not a string");
+        }
         _id = new ObjectId(_id);
         const db = await MongoConnectionManager.getConnection();
         return new Promise((resolve, reject) => {
